requirements.txt

reflex==0.8.15a1
google-genai